+++ crates/miden-objects/src/account/delta/vault.rs
@@ -5,12 +5,13 @@ use alloc::{
 };
 
 use super::{
-    AccountDeltaError, ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable,
+    AccountDeltaError, ByteReader, ByteWriter, Deserializable, DeserializationError,
+    LexicographicWord, Serializable,
 };
 use crate::{
     Felt, ONE, Word, ZERO,
-    account::{AccountId, AccountType, delta::LexicographicWord},
-    asset::{Asset, AssetVault, FungibleAsset, NonFungibleAsset},
+    account::{AccountId, AccountType},
+    asset::{Asset, FungibleAsset, NonFungibleAsset},
 };
 
 // ACCOUNT VAULT DELTA
@@ -156,35 +157,6 @@ impl AccountVaultDelta {
             .chain(self.non_fungible.filter_by_action(NonFungibleDeltaAction::Remove).map(|key| {
                 Asset::NonFungible(unsafe { NonFungibleAsset::new_unchecked(key.into()) })
             }))
-    }
-}
-
-impl From<&AssetVault> for AccountVaultDelta {
-    fn from(vault: &AssetVault) -> Self {
-        let mut fungible = BTreeMap::new();
-        let mut non_fungible = BTreeMap::new();
-
-        for asset in vault.assets() {
-            match asset {
-                Asset::Fungible(asset) => {
-                    fungible.insert(
-                        asset.faucet_id(),
-                        asset
-                            .amount()
-                            .try_into()
-                            .expect("asset amount should be at most i64::MAX by construction"),
-                    );
-                },
-                Asset::NonFungible(asset) => {
-                    non_fungible.insert(LexicographicWord::new(asset), NonFungibleDeltaAction::Add);
-                },
-            }
-        }
-
-        Self {
-            fungible: FungibleAssetDelta(fungible),
-            non_fungible: NonFungibleAssetDelta::new(non_fungible),
-        }
     }
 }